Added latest option

+++ src/Commands/ImportDump.php
@@ -2,12 +2,12 @@
 
 namespace Cybex\Protector\Commands;
 
-use Cybex\Protector\Protector;
 use Exception;
 use Illuminate\Console\Command;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\File;
+use League\Flysystem\FileNotFoundException;
 use LogicException;
 
 /**
@@ -29,15 +29,19 @@ class ImportDump extends Command
                 {--allow-production : Enable importing SQL dumps on a production system. }
                 {--force : Forces the import of the given file or remote download. Requires the dump, file or remote option. }
                 {--i|ignore-connection-filter : Ignores filter of dumps to defined connections. }
-                {--r|remote : Pull a fresh dump from the remote server as configured in the .env file. }
-                {--flush : Delete all existing dumps in the dump folder when using a remote dump}';
+                {--r|remote : Pull a fresh dump from the remote server as configured in the .env file. Will be used as fallback when combined with other options. }
+                {--flush : Delete all existing dumps in the dump folder when using a remote dump. }
+                {--l|latest : Import the most recent dump available in the configured dumps directory. }';
 
     /**
      * The console command description.
      *
      * @var string
      */
-    protected $description = 'Imports a dump.';
+    protected $description = 'Imports a local or remote database dump.';
+
+    protected const DOWNLOAD_REMOTE_DUMP       = 'Download remote dump';
+    protected const IMPORT_EXISTING_LOCAL_DUMP = 'Import existing local dump';
 
     /**
      * Create a new command instance.
@@ -60,6 +64,7 @@ class ImportDump extends Command
         $optionFile   = $this->option('file');
         $optionRemote = $this->option('remote');
         $optionForce  = $this->option('force');
+        $optionLatest = $this->option('latest');
 
         if ($optionForce && !($optionRemote || $optionFile || $optionDump)) {
             $this->error('The force option requires either the file, dump or remote option to be set.');
@@ -72,6 +77,7 @@ class ImportDump extends Command
         $basePath                = config('protector.baseDirectory');
         $destinationFilename     = $optionDump ?: $protector->createFilename();
         $relativePath            = $optionFile ?: $basePath . DIRECTORY_SEPARATOR . $destinationFilename;
+        $importFilePath          = null;
 
         $connectionName = null;
 
@@ -92,18 +98,40 @@ class ImportDump extends Command
             return;
         }
 
-        if (!($optionRemote || $optionFile || $optionDump)) {
+        if (!($optionRemote || $optionFile || $optionDump || $optionLatest)) {
             if ($this->choice('Do you want to download and import a fresh dump from the server or an existing local dump?',
                     [
-                        '1' => 'Download remote dump',
-                        '2' => 'Import existing local dump',
+                        1 => static::DOWNLOAD_REMOTE_DUMP,
+                        2 => static::IMPORT_EXISTING_LOCAL_DUMP,
                     ],
-                    'Download remote dump') == 'Download remote dump') {
+                    1) === static::DOWNLOAD_REMOTE_DUMP) {
                 $optionRemote = true;
-            };
+            }
         }
 
-        if ($optionRemote) {
+        if ($optionLatest) {
+            try {
+                $importFilePath = $protector->getLatestDumpName();
+            } catch (FileNotFoundException $fileNotFoundException) {
+                if (!$optionRemote) {
+                    $this->error($fileNotFoundException->getMessage());
+                    return;
+                } else {
+                    $this->warn(sprintf('There are no files in %s', $disk->path($basePath)));
+                }
+            }
+        } elseif ($optionFile || $optionDump) {
+            if ($disk->exists($relativePath)) {
+                $importFilePath = $relativePath;
+            } elseif (!$optionRemote) {
+                $this->error((new FileNotFoundException($relativePath))->getMessage());
+                return;
+            } else {
+                $this->warn(sprintf('File not found: %s', $relativePath));
+            }
+        }
+
+        if (!$importFilePath && $optionRemote) {
             if ($this->option('flush')) {
                 $disk->delete($disk->files($basePath));
                 $this->warn(sprintf('Deleted all files in %s', $disk->path($basePath)));
@@ -121,9 +149,9 @@ class ImportDump extends Command
             $this->line(sprintf('>>> Successfully retrieved remote dump from %s', app('protector')->getServerUrl()));
 
             $importFilePath = $fullRemoteDumpFileName;
-        } elseif ($optionFile || $optionDump) {
-            $importFilePath = $relativePath;
-        } else {
+        }
+
+        if (!$importFilePath) {
             $directoryFiles = $disk->files($basePath);
             $matchingFiles  = collect();
